requirements.txt
chromadb
openpyxl
pandas
numpy
requests
streamlit
python-dotenv
